[TIMOB-23790] Fixed bug when stopping log relaying.
Also fixed bug where an exception was not being caught. It needed to be caught and emitted as a V8 exception.

The exception was due to failing to start a service which in turn was because it detected the device over both USB and Wifi. Added code to ignore Wifi devices since we can't start services over Wifi.

Fixed bug with disconnecting from a device that has had more than one active connections.