build: add back copying of MobileDevice framework (#65)
* build: add back copying of MobileDevice framework

The MobileDevice framework can live in two locations. From Xcode 9 the first location is blocked
from being linked to which will cause build failures, the second location is not blocked  but only
seems to exist in specific circumstances (current theory is that it is where updated versions are)
placed. This fixes the build process so that we copy the framework to the build directory from one
of those locations allowing us to link against it with no issues.

* ci: lets try enabling it