build: move to using prebuildify
This replace the node-pre-gyp setup with prebuildify which will build the binaries before publish
and include them in the npm package, replacing the need for the S3 bucket that currently serves the
binaries.

This does unfortunately have the side effect that if a node version is changed and the binary does
not exist locally, then it will no longer fallback to building a fresh binary. I assume this is
probably something we could create our own package for if required, it's just that it does not
exist today.