Add support for Node.js 18 and 19
As node-pre-gyp is no longer maintained this is the path of least resistance to adding new versions,
we include a patch within this project that updates the abi_crosswalk file within node-pre-gyp that
is used to map the versions of node to their module/v8 versions. Then in run.js we call patch-package
before spawning anything to ensure that our patches are applied correctly.

This isn't by any means perfect but it's the quickest way to update the supported versions of node
without having to go and make any major refactors to the setup here, or the code in ioslib/the SDK